\name{plot.fpt}
\alias{plot.fpt}
\title{Plot FPT}
\usage{
plot.fpt(x, radii, main = "", ...)
}
\arguments{
  \item{x}{a \code{fipati} (first passage time) class
  object}

  \item{radii}{a vector of radii for x}
}
\description{
A modified version of
\code{\link[adehabitatLT]{plot.fipate}} from
\code{adehabitatLT} that plots the fpt for multiple radii.
}

